pyproject.toml
[project]
name = "transcript-to-markdown"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.1.0",
    "pyannote-audio>=3.3.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
